# api/swagger.yaml
openapi: 3.0.0
info:
  title: Personal Finance Manager API
  description: API documentation for the Personal Finance Manager application
  version: 1.0.0
  contact:
    name: Backend Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: double
        description:
          type: string
        date:
          type: string
          format: date-time
        type:
          type: string
          enum: [income, expense, transfer]
        status:
          type: string
          enum: [pending, completed, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/transactions:
    get:
      summary: Get user transactions
      tags: [Transactions]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: category_id
          schema:
            type: string
            format: uuid
        - in: query
          name: account_id
          schema:
            type: string
            format: uuid
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense, transfer]
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      summary: Create a new transaction
      tags: [Transactions]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - account_id
                - amount
                - description
                - date
                - type
              properties:
                account_id:
                  type: string
                  format: uuid
                category_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: double
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                type:
                  type: string
                  enum: [income, expense, transfer]
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
...